Cleaned up the project
Moved a bunch of files around to make things easier to understand,
renamed some classes, deleted some files that weren't being used, got
rid of a bunch of using statements that weren't being used, added a
couple of new classes.

When you run the project I've tied the incrimentDay method to the update
loop of the game so days go by really quickly, that's just to test
things like how long it takes for a seed to germinate and what not.

If you put a breakpoint in the plantedSeedObject.cs on the
Game_DayIncrimented method it should illustrate what I mean.